Update upload code to work correctly with AWS 2.1.1 Expect header change.  With the introduction of AWS sdk 2.1.1 two of the unit tests are broken.
Since we are calling inStream.pipe(multipartETag) this puts the stream into flow-mode where data will being flowing on the next tick. With files larger than 1MB AWS waits for an Expect response from the server before calling .pipe on the body stream. With the added latency most (if not all) the data is already read from the inStream pipe by the time AWS gets around to reading from it.

The cleanest fix would seem to be changing the MultipartETag class to be a pass-through stream (it is already 99.9% like this) and then setting this stream in the s3Params.body property. Another fix would be to call .pause() on the inStream but that relies on implicit .pipe() behavior turning the stream back into flow-mode and on AWS implementation details (that they are currently calling .pipe()).